WIP: mesa: Treat alpha as unorm8 for sRGB ASTC decodes
The documentation on ASTC states that a vector of UNORM8 values are
returned when decoding sRGB ASTC formats. For the alpha channel however,
it doesn't seem to specify how to get there from the UNORM16 produced after interpolation (or returned from a void-extent block).

The ASTC decoder in the VK-GL-CTS project treats the alpha channel like
the others and simply uses the top 8-bits of the UNORM16. For better
performance, we chose to do the same.