WIP: mesa: Store the color array as an int and output 32bpp
This enables the compiler to use simd instructions.

Before this patch, the compiler:
* packs the 16bit values into a 64bit qword
* moves the qword into an 128bit sse reg
* does a packed shift left by 8 operating on each word in that reg

The compiler seemed to be getting stuck on the uint16_t casts and avoided
combining the 6 and 8 right shifts. This change resolved that.

Outputing 32bits enables the compiler to save an xmm register directly instead
of having to unpack it.