WIP: main: Constant-folding for endpoint addresses
This causes the compiler to generate fixed-offset addresses for each
partition's endpoint instead of determining the offset based on a variable
holding the partition index.

It looks like we're double fetching endpoints, but we should be fine. The first
fetches all endpoints into the cache, so further accesses should be fast. I
tried making the first fetch conditional, but that broke this optimization.