WIP: main: Pre-multiply seed by two for small block
Perf is showing that select_partition is the bottleneck in write_decoded.
Particularly when it comes to computing the variables a, b, c, d.

Instead of deciding to double the x,y,z variables in those equations on a
per-pixel basis, do this per-block by doubling the seed variables instead.