fix: defer pt allocation for TexStorage in fallbacks, do validation
I neglected to defer allocation of tObj->pt in the TexStorage case, and
we need to do validation to pull in the uncompressed/transcoded data
even though we wouldn't normally do this.

textureviews still broken, to be fixed in the next patch.

Thanks to Marc Mao for help tracking down issues with immutable textures