START ASTC TRANSCODE OPTS
This is an early version of Nanley's xehp-astc-perf branch and Ken's
dynamic-dxt changes to help with Android cloud gaming use cases.

Nanley has newer versions of the patches which should likely be used
instead in the long term, but we haven't tested them.  We are unsure
whether dynamic-dxt should land or be replaced by a different approach.