dynamic-dxt: Drop !tObj->pt assert
This can happen in a case where we change the base level multiple times,
from 0..n, then back to 0 again, with no actual changes to the data.
A pt already exists with everything there; we could probably detect this
and avoid copying, but it should be pretty uncommon outside of test
cases.

We're probably taking a different approach to dynamic DXT in the long
run, so this will at least help things limp forward for now.