glsl: Replace betsy shader #includes with %s
Also, drop the commented out includes. Shader caching is disabled if
those are found.

We could attempt to concatenate the header files instead, but we choose
this method for more flexibility. For example, a shader may want to
provide a dynamic workgroup size in a gles context where
ARB_compute_variable_group_size is unsupported.

One downside to this approach is that the glsl file extensions are no
longer true. These files are now templates.