squash! iris: store a dup() of prime_fd at bo->real.prime_fd
needs_prime_fd() is being placed aboved of
iris_bo_gem_create_from_name() because it will be used by it in the
next patch.

moving the assigment of real.prime_fd to iris_bo_mark_exported_locked()
to handle cases like bo is marked as exported in
iris_resource_create_for_buffer()/iris_resource_create_for_image()
but then iris_bo_export_dmabuf() is never called.
it need to check if prime_fd == -1 because there is cases that a
imported bo is exported, so we would leak a prime_fd.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>