intel/brw/validate: Convert non-vstride validation check to use a bit per grf
Previously this check would create a mask of the bytes used in the
grf, and then shift the mask. This worked well when there was 32 bytes
in the register because a 64-bit uint64_t could easily detect that
bytes were used in the next regiter. (The next register was the high
32-bits of the `access_mask` variable.)

With Xe2, the register size becomes 64 bytes, meaning this strategy
doesn't work. Instead, let's just calculate the grf where the start
and end bytes of the field are located and then store a bit of
"access" per grf. This will be easy to extend for Xe2.

I verified this with
dEQP-VK.subgroups.arithmetic.compute.subgroupexclusivemul_u64vec4_requiredsubgroupsize
on Xe2, which otherwise would cause the program to fail to validate
because it assumed a grf was 32 bytes.

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>