iris: Add support for compressed images allocation in Xe2
Xe2 replaces auxiliary surface mapping by software to compress buffers,
instead it reserves part of the memory for the compression purpose.

To enable compress in Xe2 it is necessary bind memory with one of
the PAT indexes that has compression enabled.

So here creating 2 new iris_heap of for integrated and other for
discrete GPUs to enabled compression.
If compression requirements are met by image,
iris_resource_image_supports_pat_compression() needs to return true
so BO_ALLOC_COMPRESSED is set and then converted to one the compressed
heaps.

With this 2 new iris_heaps we gain cache and sub-allocation for free.

At this moment iris_resource_image_supports_pat_compression() is still
always returning false as there is more work needed to be done but
the foundation for the compressed allocation is here.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>